+++ crates/core/toolchain/src/tools/mod.rs
@@ -1,4 +1,4 @@
 pub mod node;
-// mod node_new;
 pub mod npm;
 pub mod pnpm;
+pub mod yarn;
