+++ crates/core/toolchain/src/tools/yarn.rs
@@ -2,9 +2,9 @@ use crate::get_path_env_var;
 use crate::tools::node::NodeTool;
 use crate::{errors::ToolchainError, DependencyManager, RuntimeTool};
 use async_trait::async_trait;
-use moon_config::PnpmConfig;
+use moon_config::YarnConfig;
 use moon_lang::LockfileDependencyVersions;
-use moon_node_lang::{pnpm, PNPM};
+use moon_node_lang::{yarn, YARN};
 use moon_utils::process::Command;
 use moon_utils::{fs, is_ci};
 use probe_core::{Executable, Probe, Resolvable, Tool};
@@ -14,19 +14,19 @@ use std::env;
 use std::path::Path;
 
 #[derive(Debug)]
-pub struct PnpmTool {
-    pub config: PnpmConfig,
+pub struct YarnTool {
+    pub config: YarnConfig,
 
     tool: NodeDependencyManager,
 }
 
-impl PnpmTool {
-    pub fn new(probe: &Probe, config: &Option<PnpmConfig>) -> Result<PnpmTool, ToolchainError> {
-        Ok(PnpmTool {
+impl YarnTool {
+    pub fn new(probe: &Probe, config: &Option<YarnConfig>) -> Result<YarnTool, ToolchainError> {
+        Ok(YarnTool {
             config: config.to_owned().unwrap_or_default(),
             tool: NodeDependencyManager::new(
                 probe,
-                probe_node::NodeDependencyManagerType::Pnpm,
+                probe_node::NodeDependencyManagerType::Yarn,
                 match &config {
                     Some(cfg) => Some(&cfg.version),
                     None => None,
@@ -34,10 +34,14 @@ impl PnpmTool {
             ),
         })
     }
+
+    fn is_v1(&self) -> bool {
+        self.config.version.starts_with('1')
+    }
 }
 
 #[async_trait]
-impl RuntimeTool for PnpmTool {
+impl RuntimeTool for YarnTool {
     fn get_version(&self) -> &str {
         self.tool.get_resolved_version()
     }
@@ -60,7 +64,7 @@ impl RuntimeTool for PnpmTool {
 }
 
 #[async_trait]
-impl DependencyManager<NodeTool> for PnpmTool {
+impl DependencyManager<NodeTool> for YarnTool {
     fn create_command(&self, node: &NodeTool) -> Command {
         let bin_path = self.tool.get_bin_path();
 
@@ -72,39 +76,51 @@ impl DependencyManager<NodeTool> for PnpmTool {
 
     async fn dedupe_dependencies(
         &self,
-        _node: &NodeTool,
-        _working_dir: &Path,
-        _log: bool,
+        node: &NodeTool,
+        working_dir: &Path,
+        log: bool,
     ) -> Result<(), ToolchainError> {
-        // pnpm doesn't support deduping, but maybe prune is good here?
-        // https://pnpm.io/cli/prune
-        // self.create_command(node)
-        //     .arg("prune")
-        //     .cwd(working_dir)
-        //     .log_running_command(log)
-        //     .exec_capture_output()
-        //     .await?;
+        // Yarn v1 doesnt dedupe natively, so use:
+        // npx yarn-deduplicate yarn.lock
+        if self.is_v1() {
+            // Will error if the lockfile does not exist!
+            if working_dir.join(self.get_lock_filename()).exists() {
+                node.exec_package(
+                    "yarn-deduplicate",
+                    vec!["yarn-deduplicate", YARN.lock_filename],
+                    working_dir,
+                )
+                .await?;
+            }
+        } else {
+            self.create_command(node)
+                .arg("dedupe")
+                .cwd(working_dir)
+                .log_running_command(log)
+                .exec_capture_output()
+                .await?;
+        }
 
         Ok(())
     }
 
     fn get_lock_filename(&self) -> String {
-        String::from(PNPM.lock_filename)
+        String::from(YARN.lock_filename)
     }
 
     fn get_manifest_filename(&self) -> String {
-        String::from(PNPM.manifest_filename)
+        String::from(YARN.manifest_filename)
     }
 
     async fn get_resolved_dependencies(
         &self,
         project_root: &Path,
     ) -> Result<LockfileDependencyVersions, ToolchainError> {
-        let Some(lockfile_path) = fs::find_upwards(PNPM.lock_filename, project_root) else {
+        let Some(lockfile_path) = fs::find_upwards(YARN.lock_filename, project_root) else {
             return Ok(FxHashMap::default());
         };
 
-        Ok(pnpm::load_lockfile_dependencies(lockfile_path)?)
+        Ok(yarn::load_lockfile_dependencies(lockfile_path).await?)
     }
 
     async fn install_dependencies(
@@ -115,12 +131,17 @@ impl DependencyManager<NodeTool> for PnpmTool {
     ) -> Result<(), ToolchainError> {
         let mut args = vec!["install"];
 
-        if is_ci() {
-            let lockfile = working_dir.join(self.get_lock_filename());
+        if self.is_v1() {
+            args.push("--ignore-engines");
+        }
 
-            // Will fail with "Headless installation requires a pnpm-lock.yaml file"
-            if lockfile.exists() {
+        if is_ci() {
+            if self.is_v1() {
+                args.push("--check-files");
                 args.push("--frozen-lockfile");
+                args.push("--non-interactive");
+            } else {
+                args.push("--immutable");
             }
         }
 
@@ -144,21 +165,23 @@ impl DependencyManager<NodeTool> for PnpmTool {
         production_only: bool,
     ) -> Result<(), ToolchainError> {
         let mut cmd = self.create_command(node);
-        cmd.arg("install");
+
+        if self.is_v1() {
+            cmd.arg("install");
+        } else {
+            cmd.args(["workspaces", "focus"]);
+            cmd.args(packages);
+
+            let workspace_plugin =
+                get_workspace_root().join(".yarn/plugins/@yarnpkg/plugin-workspace-tools.cjs");
+
+            if !workspace_plugin.exists() {
+                return Err(ToolchainError::RequiresYarnWorkspacesPlugin);
+            }
+        };
 
         if production_only {
-            cmd.arg("--prod");
-        }
-
-        for package in packages {
-            cmd.arg(if production_only {
-                "--filter-prod"
-            } else {
-                "--filter"
-            });
-
-            // https://pnpm.io/filtering#--filter-package_name-1
-            cmd.arg(format!("{}...", package));
+            cmd.arg("--production");
         }
 
         cmd.exec_stream_output().await?;
